BUG: check for file existence

+++ Testing/elx_compare_overlap.py
@@ -72,7 +72,7 @@ def main():
   seg_defm = os.path.join( options.directory, "segmentation_deformed.mha" );
   subprocess.call( [ "transformix", "-in", options.mseg, "-out", options.directory, "-tp", tpFileName ],
     stdout=subprocess.PIPE );
-  os.remove( seg_defm );
+  if( os.path.exists( seg_defm ) ) : os.remove( seg_defm );
   shutil.move( seg, seg_defm );
 
   #
@@ -94,7 +94,7 @@ def main():
   seg_defb = os.path.join( options.directory, "segmentation_baseline.mha" );
   subprocess.call( [ "transformix", "-in", options.mseg, "-out", options.directory, "-tp", tpFileName_b ],
     stdout=subprocess.PIPE );
-  os.remove( seg_defb );
+  if( os.path.exists( seg_defm ) ) : os.remove( seg_defb );
   shutil.move( seg, seg_defb );
 
   # Compute the overlap between baseline segmentation and deformed moving segmentation
